backend/requirements.txt
fastapi>=0.111.0
uvicorn[standard]>=0.30.0
langgraph>=0.1.0
langchain-groq>=0.1.0
langchain-core>=0.2.0
networkx>=3.3
pdfplumber>=0.11.0
python-dotenv>=1.0.0
python-multipart>=0.0.9
pydantic>=2.7.0
pytest>=8.0.0
httpx>=0.27.0
